dbt_debug: INFO_PRINT(..) -> WORDY_PRINT(..) cond DBTEnv::VERBOSE, add uncond INFO_PRINT(..)
Turns out INFO_PRINT(..) is confusing, as it was not clear that it is conditional on DBTEnv::VERBOSE.

Hence rename it to the short English from for Latin 'verbosus',
and hence hinting it is conditional on DBTEnv::VERBOSE.
  INFO_PRINT(..) renamed to WORDY_PRINT(..), using prefix 'Wordy:' (Verbose is just too long)

Add unconditional INFO_PRINT(..) for plain informal messages.

Hope this clears up the confusion.
Hint: I was looking for an INFO_PRINT(..) ouput myself ;-)