Fix cow_vector's custom mutable write operation, use: get_write_mutex(), copy_store() and set_store(..)
Earlier we abused get_snapshot() for the custom write operation,
leading the read-operations using the same snapshot ending in a mutated vector -> BUG.

CoW write operations always need to operate on an actual copy of the vector,
like push_back() and then replace the store_ref at the end.

Hence added copy_store() to return a copy
and refined set_store(..) as a move operator requiring std::move(..)
to avoid a copy when replacing the store_ref.