Rename jaucpp -> jaulib (internally and api-doc); Don't distributed binaries (shared lib)
As I prepare to add more languages to this utility project, rename from jaucpp -> jaulib.

Also do not distribute a binarie, i.e. no installation of a share lib artifact.
This to emphasize the intention to have artifacts being included in the
user project directly, instead of linking.
This may change, but for now it serves for best efficiency
and removes binary distribution issues with the user's project.