ringbuffer: Move the 'T nullelem' to ctor, simplifying ringbuffer declaration
Having the `nullelem` passed as reference as a template parameter
gives user the burden to have a static instance setup before declaration.
This is not nice.

For now, we better pass `nullelem` via the constructor.

We would like to pass `T nullelem` as a non-type template parameter of type `T`, a potential Class.
However, this is only allowed in C++20 and we use C++17 for now.