ringbuffer: Remove locking mutex before notify_all, leading to pessimistic re-block of notified wait() thread.
Holding the same lock @ notify_all as the waiting thread, would lead to waking up the waiting thread,
which only would be blocked again as it attempts to re-acquire the lock.

Hence removing it.

A lock is also not required for the readPos or writePos field,
as it is atomic of memory_order_seq_cst.