file_util: Add traverse_event as path_visitor() argument and traverse_options - both supporting visit(); remove() uses traverse_options
It is helpful to have path_visitor() be informed about the event it is called,
i.e. a symlink and/or file, at directory-entry or -exit or when processing a non-followed symlink to a directory.

traverse_options being passed to visit() renders it more versatile in conjuction with traverse_event,
especially for dir_entry and dir_exit cases where potentially different actions are performed
also allowing the caller to specify which call shall be made, none or both.

remove() uses traverse_options, allowing same visit() to be applied.
Only dir_exit is explicit set and dir_entry cleared due to its M.O.