file_stats: Support new fmode_t types sock, blk, chr and fifo, move fd from type -> attribute, support sole file_stats(fd) ctor, ...
A file descriptor (fd) is not a fmode_t type, but an attribute clarifying if instance
was constructed via a fd w/o a path (including named fd) or not.

Hence is_fd() -> has_fd(), useful to avoid ::remove() a file.

For example, stdin could be either a file, a chr (terminal) or a fifo (pipe)
where all variants are has_fd() == true and size() == 0.

+++

Allow instantiation via a sole fd, as required for upcoming pure fd handling.

Taking benefit of the POSIX openat() function group using AT_EMPTY_PATH
and an empty path.

Note that readlinkat() just uses an empty path as the signal for using dirfd alone.

+++

fix named pipe handling: If detecting a named-fd (could be still a link type as well),
simply use the resulting fd and open it via stat().

This avoid issues:
  - on FreeBSD where opening the named-fd itself gives the wrong fmode_t
   (always chr instead of fifo)

  - on Linux where the final fifo/pipe file can't be opened.