jau::codec::base::decode*(): Validate alphabet::code_point() result with used base, since used base <= alphabet::max_base(); Same for Java BaseCodec.decode*()
It is designed to use an alphabet with a max_base > currently being used base,
to reduce the number of alphabet specifications.

This design might change in favor of using a fixed base alphabet,
in which case the variable base argument will be dropped as well.

However, as long base <= max_base,
we must validate that alphabet::code_point() < base [0..base).
Otherwise values exceeding the desired code could have been introduced by an attacker.