Robustness and versatility of byte_util: Use lb_endian enum class instead of 'boolean isLittleEndian'; Remove 'nsize_t byte_offset' and rely on given pointer
Using a byte offset of type nsize_t would potentially reduce addressable space on a given platform
and is not required for C++ as the pointer arithmetic can be used by the caller.

Dropping the byte offset exposed the weakness of using a bool type for the little- big-endian switch,
as an integer is automatically converted to bool.
Hence we now use an enum class lb_endian only denoting little-, big-endian
and additional native (the actual platform's encoding).